LayoutTests/imported/w3c/web-platform-tests/navigation-api/scroll-behavior/after-transition-explicit-scroll-expected.txt


FAIL scroll: scroll() should preempt after-transition assert_equals: expected 0 but got 431
